Add files via upload
the above code is a java code for binary number. The binary number system is base-2, meaning it uses two digits: 0 and 1.Binary numbers in Java are essential for efficient computation and form the backbone of how computers process and store information. Understanding this theory provides a solid foundation for exploring deeper topics like data structures, algorithms, and system-level programming.